fix(faq): address review feedback (stable category IDs, server component, footer link test)
- Add stable id field to FAQ categories to replace translated labels
  in id/aria-labelledby attributes (invalid HTML ids with spaces).
- Convert /faq page to a Server Component; extract the interactive
  accordion into FaqAccordion client component.
- Add Playwright test covering the FAQ footer link on the authenticated
  landing page (trip-planner welcome state).
